requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF
faiss-cpu
tiktoken